+++ src/ska_mid_dish_manager/component_managers/dish_manager_cm.py
@@ -103,6 +103,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             achievedpointing=[0.0, 0.0, 0.0],
             attenuationpolh=0.0,
             attenuationpolv=0.0,
+            dscpowerlimitkw=10.0,
             kvalue=0,
             scanid="",
             trackinterpolationmode=TrackInterpolationMode.SPLINE,
@@ -164,6 +165,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                 band4pointingmodelparams=[],
                 band5apointingmodelparams=[],
                 band5bpointingmodelparams=[],
+                dscpowerlimitkw=10.0,
                 trackinterpolationmode=TrackInterpolationMode.SPLINE,
                 actstaticoffsetvaluexel=None,
                 actstaticoffsetvalueel=None,
@@ -408,6 +410,13 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                 ds_component_state["pointingstate"],
             )
             self._update_component_state(pointingstate=ds_component_state["pointingstate"])
+
+        if "dscpowerlimitkw" in kwargs:
+            self.logger.debug(
+                ("DSC Power Limit kW changed [%s]"),
+                ds_component_state["dscpowerlimitkw"],
+            )
+            self._update_component_state(dscpowerlimitkw=ds_component_state["dscpowerlimitkw"])
 
         # spf bandInFocus
         if not self.is_device_ignored("SPF") and (
@@ -1176,6 +1185,21 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             )
 
         return (ResultCode.OK, "Successfully updated pseudoRandomNoiseDiodePars on SPFRx")
+
+    def set_dsc_power_limit_kw(
+        self,
+        power_limit,
+    ) -> None:
+        """Set the DSC Power Limit kW on the DS."""
+        self._update_component_state(dscpowerlimitkw=power_limit)
+        ds_cm = self.sub_component_managers["DS"]
+        try:
+            ds_cm.write_attribute_value("dscPowerLimitKw", power_limit)
+            self.logger.debug("Successfully updated dscPowerLimitKw on DS.")
+        except (LostConnection, tango.DevFailed):
+            self.logger.error("Failed to update dscPowerLimitKw on DS.")
+            raise
+        return (ResultCode.OK, "Successfully updated dscPowerLimitKw on DS")
 
     def _get_device_attribute_property_value(self, attribute_name) -> Optional[str]:
         """